requirements.txt
kafka-python==1.4.7
Django==3.0.6
psycopg2==2.8.4
django-model-utils==4.0.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
git+https://github.com/project-a/herbie-json-schema.git
jsonschema==3.2.0
django-environ==0.4.5
social-auth-app-django==3.1.0
social-auth-core==3.2.0
pytest==5.3.1
pytest-django==3.5.1
google-cloud-pubsub==1.1.0

